airflow_automation/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pyyaml
beautifulsoup4
selenium
pandas 
numpy
webdriver-manager
requests
